v1.1.18

+++ fix-webpack-output.js
@@ -13,14 +13,24 @@ async function fixWebpackOutput() {
   const sourceMain = path.join(webpackDir, archFolder, 'main')
   const targetMain = path.join(webpackDir, 'main')
 
+  const sourceRenderer = path.join(webpackDir, archFolder, 'renderer')
+  const targetRenderer = path.join(webpackDir, 'renderer')
+
   if (!(await fs.pathExists(sourceMain))) {
     console.error(`Pasta de origem não encontrada: ${sourceMain}`)
     process.exit(1)
   }
 
+  if (!(await fs.pathExists(sourceRenderer))) {
+    console.error(`Pasta de origem não encontrada: ${sourceRenderer}`)
+    process.exit(1)
+  }
+
   await fs.remove(targetMain)
+  await fs.remove(targetRenderer)
 
   await fs.copy(sourceMain, targetMain)
+  await fs.copy(sourceRenderer, targetRenderer)
   console.log(`Arquivos copiados de ${sourceMain} para ${targetMain}`)
 }
